+++ client/plyer/components/Modal/UpdateProposalModal.tsx
@@ -20,7 +20,7 @@ export const UpdateProposalModal = ({data, opened, ...props}: Props) => {
         form.setValues({
             content: data?.content
         })
-    }, [opened, data])
+    }, [opened, data, form])
 
     const onSubmit = ({content }: any) => {
         updateReques({