+++ client/plyer/components/Modal/UpdateRequestModal.tsx
@@ -18,7 +18,7 @@ export const UpdateRequestModal = ({data, opened, ...props}: Props) => {
         form.setValues({
             content: data?.content
         })
-    }, [opened, data])
+    }, [opened, data, form])
 
     const onSubmit = ({content}: any) => {
         updateReques({