+++ client/super-admin/components/Table/DepartementTable.tsx
@@ -189,8 +189,9 @@ const DepartementTable = ({ data, searchValue, onEditModal, ...props }: Props) =
   ));
 
   React.useEffect(() => {
-    handleSearchChange(searchValue);
-  }, [searchValue]);
+    setSearch(searchValue);
+    setSortedData(sortData(data, { sortBy, reversed: reverseSortDirection, search: searchValue }));
+  }, [searchValue, data, sortBy, reverseSortDirection]);
 
   return (
     <Table