+++ client/super-admin/components/Table/EmployeerSettingsTable.tsx
@@ -238,8 +238,9 @@ const EmployeerSettingsTable = ({ data, searchValue, ...props }: Props) => {
   ));
 
   React.useEffect(() => {
-    handleSearchChange(searchValue);
-  }, [searchValue]);
+    setSearch(searchValue);
+    setSortedData(sortData(data, { sortBy, reversed: reverseSortDirection, search: searchValue }));
+  }, [searchValue, data, sortBy, reverseSortDirection]);
 
   return (
     <Table