+++ client/super-admin/components/Table/ExpenseEmployeeTable.tsx
@@ -186,8 +186,9 @@ const ExpenseEmployeeTable = ({ data, searchValue, onEditModal, ...props }: Prop
   ));
 
   React.useEffect(() => {
-    handleSearchChange(searchValue);
-  }, [searchValue]);
+    setSearch(searchValue);
+    setSortedData(sortData(data, { sortBy, reversed: reverseSortDirection, search: searchValue }));
+  }, [searchValue, data, sortBy, reverseSortDirection]);
 
   return (
     <Table