+++ client/super-admin/components/Table/RequestEmployeeTable.tsx
@@ -222,8 +222,9 @@ const RequestEmployeeTable = ({ data, searchValue, onEditModal, ...props }: Prop
   ));
 
   React.useEffect(() => {
-    handleSearchChange(searchValue);
-  }, [searchValue]);
+    setSearch(searchValue);
+    setSortedData(sortData(data, { sortBy, reversed: reverseSortDirection, search: searchValue }));
+  }, [searchValue, data, sortBy, reverseSortDirection]);
 
   return (
     <Table