+++ client/super-admin/components/Table/RequestTable.tsx
@@ -235,8 +235,9 @@ const RequestTable = ({ data, searchValue, onEditModal, ...props }: Props) => {
   ));
 
   React.useEffect(() => {
-    handleSearchChange(searchValue);
-  }, [searchValue]);
+    setSearch(searchValue);
+    setSortedData(sortData(data, { sortBy, reversed: reverseSortDirection, search: searchValue }));
+  }, [searchValue, data, sortBy, reverseSortDirection]);
 
   return (
     <Table