+++ client/super-admin/components/Table/SchedulesTable.tsx
@@ -159,8 +159,9 @@ const SchedulesTable = ({ data, searchValue, ...props }: Props) => {
   ));
 
   React.useEffect(() => {
-    handleSearchChange(searchValue);
-  }, [searchValue]);
+    setSearch(searchValue);
+    setSortedData(sortData(data, { sortBy, reversed: reverseSortDirection, search: searchValue }));
+  }, [searchValue, data, sortBy, reverseSortDirection]);
 
   return (
     <Table